CoNLI/CoNLI/requirements.txt
rouge_score
sacrebleu
openai
azure.identity
scikit-learn
transformers
torch
spacy
pandas
nltk
hydra-core
azure-keyvault-secrets
azure-ai-textanalytics